Clang-tidy

+++ esphome/components/lvgl/lvgl_esphome.cpp
@@ -10,7 +10,7 @@ unsigned long lv_millis(void) { return esphome::millis(); }
 
 static unsigned cap_bits = MALLOC_CAP_SPIRAM | MALLOC_CAP_8BIT;
 
-void *lv_custom_mem_alloc(unsigned int size) {
+void *lv_custom_mem_alloc(size_t size) {
   void *ptr;
   ptr = heap_caps_malloc(size, cap_bits);
   if (ptr == nullptr) {
@@ -36,7 +36,7 @@ void lv_custom_mem_free(void *ptr) {
   heap_caps_free(ptr);
 }
 
-void *lv_custom_mem_realloc(void *ptr, unsigned int size) {
+void *lv_custom_mem_realloc(void *ptr, size_t size) {
 #ifdef ESPHOME_LOG_HAS_VERBOSE
   esphome::esph_log_v(esphome::lvgl::TAG, "realloc %p: %u", ptr, size);
 #endif
